Install Widgy with python 3 fixes
This follows the direction in the Widgy Quickstart Guide:
http://docs.wid.gy/en/latest/tutorials/widgy-mezzanine-tutorial.html

Many fixes were needed just to make it install and run in Python 3.4.
The file requirements.txt contains many bleeding edge repositories for
various dependencies, and refers to a customized version of Widgy with a
few fixes that are not incorporated in the official repository yet as of
this commit.

Many bugs remain: don't expect a functional system. It does install, run
and show the admin interface.